backend/requirements.txt
fastapi
uvicorn
pydantic-settings
google-generativeai
python-dotenv
httpx
python-docx
